RPET-746: Update Pact test for PBA interaction with CaseType
